Refactor: pull pk3 to vegadisk
