+++ libraries/vegadisk/vsfilesystem.h
@@ -35,7 +35,7 @@
 #include <iostream>
 #include <cstdarg>
 #include "gfx_generic/vec.h"
-#include "root_generic/pk3.h"
+#include "vegadisk/pk3.h"
 #include "src/gnuhash.h"
 #include "src/vs_logging.h"
